+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 - fix: convert har to JSON format testcase
 - fix: missing ${var} handling in overriding config variables
 - fix: SyntaxError caused by quote in case of headers."Set-Cookie"
+- fix: FileExistsError when specified project name conflicts with existed file
 
 ## 3.0.12 (2020-06-14)